punjenje baze za ingridients na pocetku aplikacije

+++ app/src/main/java/rs/raf/rma/nutritiontrackerrma/data/repositories/ingredients/IngredientsRepositoryImpl.kt
@@ -32,7 +32,7 @@ class IngredientsRepositoryImpl(
 
                 val entities = items.map {
                     IngredientEntity (
-                        it.idIngredient,
+                        it.idIngredient.toInt(),
                         it.strIngredient,
                         100,
                         0.0
